+++ src/library/DIPS.Mobile.UI/API/Builder/AppHostBuilderExtensions.cs
@@ -2,6 +2,7 @@ using DIPS.Mobile.UI.API.Library;
 using DIPS.Mobile.UI.Components.Chips;
 using DIPS.Mobile.UI.Components.ContextMenus;
 using DIPS.Mobile.UI.Components.Images.NativeIcon;
+using DIPS.Mobile.UI.Components.Lists;
 using DIPS.Mobile.UI.Components.Pickers.DateAndTimePicker;
 
 #if __ANDROID__
@@ -9,10 +10,11 @@ using DIPS.Mobile.UI.Effects.Modal;
 using DIPS.Mobile.UI.Effects.Modal.Android;
 #endif
 using DIPS.Mobile.UI.Effects.Touch;
-using Microsoft.Maui.Handlers;
 using Microsoft.Maui.LifecycleEvents;
-using Button = DIPS.Mobile.UI.Components.Buttons.Button;
 using ButtonHandler = DIPS.Mobile.UI.Components.Buttons.ButtonHandler;
+using Button = DIPS.Mobile.UI.Components.Buttons.Button;
+using CollectionView = DIPS.Mobile.UI.Components.Lists.CollectionView;
+using CollectionViewHandler = DIPS.Mobile.UI.Components.Lists.CollectionViewHandler;
 using ContextMenuPlatformEffect = DIPS.Mobile.UI.Components.ContextMenus.ContextMenuPlatformEffect;
 using DatePickerHandler = DIPS.Mobile.UI.Components.Pickers.DatePicker.DatePickerHandler;
 using Image = DIPS.Mobile.UI.Components.Images.Image.Image;
@@ -20,6 +22,8 @@ using ImageButton = DIPS.Mobile.UI.Components.Images.ImageButton;
 using ImageHandler = DIPS.Mobile.UI.Components.Images.Image.ImageHandler;
 using Label = DIPS.Mobile.UI.Components.Labels.Label;
 using LabelHandler = DIPS.Mobile.UI.Components.Labels.LabelHandler;
+using ScrollView = DIPS.Mobile.UI.Components.Lists.ScrollView;
+using ScrollViewHandler = DIPS.Mobile.UI.Components.Lists.ScrollViewHandler;
 using SearchBar = DIPS.Mobile.UI.Components.Searching.SearchBar;
 using SearchBarHandler = DIPS.Mobile.UI.Components.Searching.SearchBarHandler;
 using TimePickerHandler = DIPS.Mobile.UI.Components.Pickers.TimePicker.TimePickerHandler;
@@ -51,6 +55,8 @@ public static partial class AppHostBuilderExtensions
             handlers.AddHandler<Image, ImageHandler>();
             handlers.AddHandler<Button, ButtonHandler>();
             handlers.AddHandler<Label, LabelHandler>();
+            handlers.AddHandler<CollectionView, CollectionViewHandler>();
+            handlers.AddHandler<ScrollView, ScrollViewHandler>();
             
             AddPlatformHandlers(handlers);
         });