+++ src/library/DIPS.Mobile.UI/Components/Lists/CollectionView/CollectionView.Properties.cs
@@ -22,6 +22,21 @@ public partial class CollectionView
         set => SetValue(HasAdditionalSizeAtTheEndProperty, value);
     }
 
+    public static readonly BindableProperty ShouldBounceProperty = BindableProperty.Create(
+        nameof(ShouldBounce),
+        typeof(bool),
+        typeof(CollectionView),
+        defaultValue: true);
+
+    /// <summary>
+    /// Determines if the collection view should bounce.
+    /// </summary>
+    public bool ShouldBounce
+    {
+        get => (bool)GetValue(ShouldBounceProperty);
+        set => SetValue(ShouldBounceProperty, value);
+    }
+    
     public static readonly BindableProperty HasAdditionalSizeAtTheEndProperty = BindableProperty.Create(
         nameof(HasAdditionalSpaceAtTheEnd),
         typeof(bool),