+++ src/library/DIPS.Mobile.UI/Components/Lists/CollectionView/iOS/CollectionViewHandler.cs
@@ -1,0 +1,19 @@
+using Microsoft.Maui.Platform;
+using UIKit;
+
+namespace DIPS.Mobile.UI.Components.Lists;
+
+public partial class CollectionViewHandler
+{
+    private static partial void MapShouldBounce(CollectionViewHandler handler,
+        Microsoft.Maui.Controls.CollectionView virtualView)
+    {
+        if (handler.PlatformView.Subviews[0] is UICollectionView uiCollectionView)
+        {
+            if (virtualView is CollectionView collectionView)
+            {
+                uiCollectionView.Bounces = collectionView.ShouldBounce;
+            }
+        }
+    }
+}