+++ src/library/DIPS.Mobile.UI/Components/Lists/ScrollView/ScrollView.Properties.cs
@@ -1,0 +1,19 @@
+namespace DIPS.Mobile.UI.Components.Lists;
+
+public partial class ScrollView
+{
+    public static readonly BindableProperty ShouldBounceProperty = BindableProperty.Create(
+        nameof(ShouldBounce),
+        typeof(bool),
+        typeof(CollectionView),
+        defaultValue: true);
+
+    /// <summary>
+    /// Determines if the collection view should bounce.
+    /// </summary>
+    public bool ShouldBounce
+    {
+        get => (bool)GetValue(ShouldBounceProperty);
+        set => SetValue(ShouldBounceProperty, value);
+    }
+}